Did some testing.  Fixed a number of typos.  (In a production system, unit tests would be a part of the final deliverable.)
Cleaned up syntax.

Realized my logic differed from the deliverables slightly: If I'm reading the requirements correctly, an elevator on the 30th floor heading to the first floor should answer a call on the first floor before an idling elevator on the 2nd floor.  Rather than question this deliverable, I've modified the system to deliver the desired final result.

Final commit.

I'd just like to add that I spend a lot of time thinking about this very problem.  The elevator in my building has to be about the slowest elevator in the world and I keep thinking how with fewer than a hundred lines of code it could be made far less annoying and with a couple thousand lines of code and a database it could be made a downright pleasure.

See, the elevator in my building simply sits wherever it was last ridden to.  Because it's so slow, it's always on the wrong floor and you have to wait for it.  For instance in the morning, when everyone is heading to work, everyone is heading down to the parking garage.  So the first person to leave for work in the morning presses the button and let's say the elevator happens to be at their floor already.  The doors open immediately, the person steps in, and they enjoy their nice, slow ride down to the garage.  The elevator now sits in the garage, languishing.  After a few minutes, the next person to head to work presses the button.  No matter what floor they're on, it's not going to be the garage.  So the elevator doors close excruciatingly slowly.  The elevator gathers up it's courage, and finally lurches slowly to the floors above.  It can take nearly a full minute to reach the top floor.  (There are only three stories, plus two garage levels.)  By the time it makes it to the top floor from the bottom of the garage, a queue is likely to be forming on levels two and one, forcing it to stop at every level along the way.  If it had dumbly returned to the lobby when not answering a call, it would've saved half the time of its journey.  If it were made just slightly smarter and returned to the upper floors in the morning and the garage floors in the afternoon, much anguish would be saved.  If it were just slightly smarter than that and kept a record of what times it was summoned on which floors on particular days of the week, it would be able to anticipate its next summons, making it much, much less painful to use with absolutely no upgrade to the hardware.  But that's all beyond the scope of this project.  Thanks for the fun challenge!

 -Max